MFAC Refactor (#558)
* M-FAC refactor to comply with pruning modifier refactor, along with smaller-scale updates to MFAC implementation and testing
