+++ tests/sparseml/pytorch/optim/test_mask_pruning_scorer.py
@@ -17,8 +17,6 @@ import torch
 
 from sparseml.pytorch.optim import (
     MagnitudePruningParamsScorer,
-    MFACOptions,
-    MFACPruningParamsScorer,
     MovementPruningParamsScorer,
     create_pruning_param_scorer,
 )
@@ -41,9 +39,6 @@ def _fake_params_random_update(params):
         ("magnitude", 0),
         ("magnitude", 1),
         ("movement", 5),
-        (MFACOptions(num_grads=10), 12),
-        (MFACOptions(num_grads=10, fisher_block_size=24), 10),
-        (MFACOptions(num_grads=8, num_pages=4 if torch.cuda.is_available() else 1), 8),
     ],
 )
 def test_pruning_scorer(score_type, n_updates):
@@ -75,8 +70,6 @@ def test_pruning_scorer(score_type, n_updates):
     [
         (MagnitudePruningParamsScorer, "magnitude"),
         (MovementPruningParamsScorer, "movement"),
-        (MFACPruningParamsScorer, "MFAC"),
-        (MFACPruningParamsScorer, MFACOptions()),
     ],
 )
 def test_create_pruning_param_scorer(expected_class, score_type):
