+++ tests/sparseml/pytorch/optim/test_modifier_pruning.py
@@ -23,8 +23,6 @@ from sparseml.pytorch.nn import Identity
 from sparseml.pytorch.optim import LayerPruningModifier
 from sparseml.pytorch.optim import LegacyGMPruningModifier as GMPruningModifier
 from sparseml.pytorch.optim import (
-    MFACGlobalPruningModifier,
-    MFACPruningModifier,
     MovementPruningModifier,
     StructuredPruningModifier,
     load_mask_creator,
@@ -561,110 +559,4 @@ def test_structured_pruning_yaml():
         yaml_modifier.param_groups
         == serialized_modifier.param_groups
         == obj_modifier.param_groups
-    )
-
-
-def test_mfac_pruning_yaml():
-    init_sparsity = 0.05
-    final_sparsity = 0.8
-    start_epoch = 5.0
-    end_epoch = 15.0
-    update_frequency = 1.0
-    params = "__ALL_PRUNABLE__"
-    inter_func = "cubic"
-    mask_type = "unstructured"
-    global_sparsity = False
-    mfac_options = {"num_grads": 64, "available_gpus": ["cuda:0", "cuda:1"]}
-    yaml_str = f"""
-    !MFACPruningModifier
-        init_sparsity: {init_sparsity}
-        final_sparsity: {final_sparsity}
-        start_epoch: {start_epoch}
-        end_epoch: {end_epoch}
-        update_frequency: {update_frequency}
-        params: {params}
-        inter_func: {inter_func}
-        mask_type: {mask_type}
-        global_sparsity: {global_sparsity}
-        mfac_options: {mfac_options}
-    """
-    yaml_modifier = MFACPruningModifier.load_obj(yaml_str)
-    serialized_modifier = MFACPruningModifier.load_obj(
-        str(yaml_modifier)
-    )  # type: MFACPruningModifier
-    obj_modifier = MFACPruningModifier(
-        init_sparsity=init_sparsity,
-        final_sparsity=final_sparsity,
-        start_epoch=start_epoch,
-        end_epoch=end_epoch,
-        update_frequency=update_frequency,
-        params=params,
-        inter_func=inter_func,
-        mask_type=mask_type,
-        global_sparsity=global_sparsity,
-        mfac_options=mfac_options,
-    )
-
-    assert isinstance(yaml_modifier, MFACPruningModifier)
-    _test_pruning_modifier_serialization_vals(
-        yaml_modifier, serialized_modifier, obj_modifier
-    )
-    assert (
-        str(yaml_modifier.global_sparsity)
-        == str(serialized_modifier.global_sparsity)
-        == str(obj_modifier.global_sparsity)
-    )
-    assert (
-        yaml_modifier.mfac_options
-        == serialized_modifier.mfac_options
-        == obj_modifier.mfac_options
-    )
-
-
-def test_global_mfac_pruning_yaml():
-    init_sparsity = 0.05
-    final_sparsity = 0.8
-    start_epoch = 5.0
-    end_epoch = 15.0
-    update_frequency = 1.0
-    params = "__ALL_PRUNABLE__"
-    inter_func = "cubic"
-    mask_type = "unstructured"
-    mfac_options = {"num_grads": 64, "available_gpus": ["cuda:0", "cuda:1"]}
-    yaml_str = f"""
-    !MFACGlobalPruningModifier
-        init_sparsity: {init_sparsity}
-        final_sparsity: {final_sparsity}
-        start_epoch: {start_epoch}
-        end_epoch: {end_epoch}
-        update_frequency: {update_frequency}
-        params: {params}
-        inter_func: {inter_func}
-        mask_type: {mask_type}
-        mfac_options: {mfac_options}
-    """
-    yaml_modifier = MFACGlobalPruningModifier.load_obj(yaml_str)
-    serialized_modifier = MFACGlobalPruningModifier.load_obj(
-        str(yaml_modifier)
-    )  # type: MFACGlobalPruningModifier
-    obj_modifier = MFACGlobalPruningModifier(
-        init_sparsity=init_sparsity,
-        final_sparsity=final_sparsity,
-        start_epoch=start_epoch,
-        end_epoch=end_epoch,
-        update_frequency=update_frequency,
-        params=params,
-        inter_func=inter_func,
-        mask_type=mask_type,
-        mfac_options=mfac_options,
-    )
-
-    assert isinstance(yaml_modifier, MFACGlobalPruningModifier)
-    _test_pruning_modifier_serialization_vals(
-        yaml_modifier, serialized_modifier, obj_modifier
-    )
-    assert (
-        yaml_modifier.mfac_options
-        == serialized_modifier.mfac_options
-        == obj_modifier.mfac_options
     )
